Split up test_model.py
